dom/tests/mochitest/crypto/test_no_legacy.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test lack of legacy window.crypto features when
         MOZ_DISABLE_CRYPTOLEGACY is set or dom.unsafe_legacy_crypto.enabled is false</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script class="testbody" type="text/javascript">

function test_unsafe_legacy_crypto_disabled() {
  ok("crypto" in window, "crypto in window");
  ok(!("version" in window.crypto), "version not in window.crypto");
  ok(!("enableSmartCardEvents" in window.crypto),
     "enableSmartCardEvents not in window.crypto");
  ok(!("generateCRMFRequest" in window.crypto),
     "generateCRMFRequest not in window.crypto");
  ok(!("importUserCertificates" in window.crypto),
     "importUserCertificates not in window.crypto");
  ok(!("signText" in window.crypto), "signText not in window.crypto");

  SimpleTest.finish();
}

SpecialPowers.pushPrefEnv({"set": [["dom.unsafe_legacy_crypto.enabled", false]]},
                          test_unsafe_legacy_crypto_disabled);
SimpleTest.waitForExplicitFinish();

</script>
</body></html>
